MOW-1365 Undo for ordering

See merge request ggb/geogebra!6372

+++ common/src/main/java/org/geogebra/common/io/ConsElementXMLHandler.java
@@ -1859,7 +1859,7 @@ public class ConsElementXMLHandler {
 
 	private void handleOrdering(LinkedHashMap<String, String> attrs) {
 		try {
-			geo.setOrdering(Integer.parseInt(attrs.get("val")));
+			geo.setOrdering(Double.parseDouble(attrs.get("val")));
 		} catch (RuntimeException e) {
 			// no or incorrect ordering
 		}
